requirements.txt
appdirs==1.4.3
asgiref==3.2.7
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
Markdown==3.2.1
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytesseract==0.3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tox==3.14.6
virtualenv==20.0.18
whitenoise==5.0.1
zipp==3.1.0
